Created 2019-07-02 transparency report
Per the [bylaws](https://docs.carpentries.org/topic_folders/policies/enforcement-guidelines.html#accountability): At the end of every quarter, the Executive Council will publish an aggregated count of the incidents the Code of Conduct Committee reviewed, indicating how many reports it received, how many incidents it investigated independently, how many times it acted unilaterally, and, for each of these, under which part of the Code of Conduct the incident was classified. Here is a draft of this report.